backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
groq
pytest
pytest-asyncio
ruff
mypy
httpx
